Refactor: Move location
